third_party/blink/web_tests/platform/linux-chrome/external/wpt/client-hints/accept-ch/cache-revalidation.https-expected.txt

This is a testharness.js-based test.
[FAIL] Same headers sent for revalidation request
  assert_equals: expected "vhgcfkuxxhkguwurvfgj" but got "ryryshskkhlokhyjwczw"
Harness: the test ran to completion.
